requirements.txt
Flask==1.1.1
Flask-Cors==3.0.8
Flask-WTF==0.14.2
WTForms==2.2.1
Werkzeug==0.16.0
pathlib2==2.3.3
numpy==1.17.4
spacy==2.2.2
spacy-lookups-data==0.2.0
absl-py==0.7.1
tensorflow==2.0.0
tensorflow-estimator==2.0.1
tensorflow-hub==0.7.0
pytextrank==2.0.0
textacy==0.9.1
docx2txt==0.8
